Complete migration testing, fix bugs and do prod ready build testing.
