+++ app/Routers/CustomerRoute.tsx
@@ -1,41 +1,42 @@
-import React from "react";
-import { Routes, Route } from "react-router-dom";
-import HomePage from "../customer/pages/HomePage/HomePage";
-import Cart from "../customer/components/Cart/Cart";
-import Navigation from "../customer/components/navigation/Navigation";
-import Footer from "../customer/components/Footer/Footer";
-import Product from "../customer/components/Product/Product";
-import ProductDetails from "../customer/components/ProductDetails/ProductDetails";
-import Checkout from "../customer/components/Checkout/Checkout";
-import Order from "../customer/components/Order/Order";
-import OrderDetails from "../customer/components/Order/OrderDetails";
+// import Cart from "app/components/Cart/Cart";
+// import Checkout from "app/components/Checkout/Checkout";
+// import Footer from "app/components/Footer/Footer";
+// import Navigation from "app/components/navigation/Navigation";
+// import Order from "app/components/Order/Order";
+// import OrderDetails from "app/components/Order/OrderDetails";
+// import Product from "app/components/Product/Product";
+// import ProductDetails from "app/components/ProductDetails/ProductDetails";
+// import HomePage from "app/page";
+// import React from "react";
+// import { Routes, Route } from "react-router-dom";
+// // import HomePage from "../customer/pages/HomePage/HomePage";
 
-const CustomerRoute = () => {
-  return (
-    <div>
-      <div>
-        <Navigation />
-      </div>
-      <Routes>
-        <Route path="/" element={<HomePage />}></Route>
-        <Route path="/cart" element={<Cart />}></Route>
-        <Route
-          path="/:LevelOne/:LevelTwo/:LevelThree"
-          element={<Product />}
-        ></Route>
-        <Route path="/product/:productId" element={<ProductDetails />}></Route>
-        <Route path="/checkout" element={<Checkout />}></Route>
-        <Route path="/account/order" element={<Order />}></Route>
-        <Route
-          path="/account/order/:orderId"
-          element={<OrderDetails />}
-        ></Route>
-      </Routes>
-      <div>
-        <Footer />
-      </div>
-    </div>
-  );
-};
+// const CustomerRoute = () => {
+//   return (
+//     <div>
+//       <div>
+//         <Navigation />
+//       </div>
+//       <Routes>
+//         <Route path="/" element={<HomePage />}></Route>
+//         <Route path="/cart" element={<Cart />}></Route>
+//         <Route
+//           path="/:LevelOne/:LevelTwo/:LevelThree"
+//           element={<Product />}
+//         ></Route>
+//         <Route path="/product/:productId" element={<ProductDetails />}></Route>
+//         <Route path="/checkout" element={<Checkout />}></Route>
+//         <Route path="/account/order" element={<Order />}></Route>
+//         <Route
+//           path="/account/order/:orderId"
+//           element={<OrderDetails />}
+//         ></Route>
+//       </Routes>
+//       <div>
+//         <Footer />
+//       </div>
+//     </div>
+//   );
+// };
 
-export default CustomerRoute;
+// export default CustomerRoute;
