+++ app/components/Cart/CartItem.tsx
