+++ app/components/Checkout/Checkout.tsx
@@ -1,10 +1,11 @@
+'use client'
 import Box from '@mui/material/Box';
 import Stepper from '@mui/material/Stepper';
 import Step from '@mui/material/Step';
 import StepLabel from '@mui/material/StepLabel';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
-import { useLocation } from 'react-router-dom';
+
 import DeliveryAddressForm from './DeliveryAddressForm';
 import OrderSummary from './OrderSummary';
 import { useState } from 'react';
@@ -13,8 +14,6 @@ const steps = ['Login', 'Add Delivery Address', 'Order Summary', 'Payment'];
 
 export default function Checkout() {
   const [activeStep, setActiveStep] = useState(0);
-  const location = useLocation();
-  const querySearch = new URLSearchParams(location.search)
 
   let step = 0;
