+++ app/components/HomeSectionCarousel/HomeSectionCarousel.tsx
