+++ app/components/navigation/brand.png
