+++ next.config.mjs
@@ -2,6 +2,9 @@
 const nextConfig = {
   output: 'export', // Outputs a Single-Page Application (SPA).
   distDir: './dist', // Changes the build output directory to `./dist/`.
+  images:{
+    unoptimized: true
+  }
 }
  
 export default nextConfig
