+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "major",
-  "version": "0.1.0",
+  "name": "shopsavvy",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.11.1",
@@ -16,12 +16,11 @@
     "@types/node": "^20.14.10",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
+    "autoprefixer": "^10.4.19",
     "next": "^14.2.4",
+    "postcss": "^8.4.39",
     "react": "^18.2.0",
     "react-alice-carousel": "^2.8.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.20.1",
-    "react-scripts": "5.0.1",
     "typescript": "^5.5.3",
     "web-vitals": "^2.1.4"
   },
